Better handling of changes in the number of installed CPUs
This patch aims to cover the following use cases:

 - The user upgraded the cpufreq plugin to a newer version
 - The user changed the number of CPU cores or SMT (hyper-threading)
   in the BIOS
 - The user moved HDD/SSD to a different machine
 - The user downgraded the CPU in the machine

This is implemented by making CPU_MIN, CPU_AVG and CPU_MAX
independent from the number of CPUs in the machine.

On the 1st time an invalid CPU number setting is detected,
the plugin will display a warning message. This warning
message will likely be displayed to most cpufreq users
after upgrading cpufreq to version 1.2.3+.