Create the label widget if it does not exist
This fixes the following bug: If (show_icon, show_freq, show_governor) is
initially (true,false,false) then changing it to (*,true,*) or (*,*,true)
via the configuration dialog does not display the label.

This patch is slightly larger because it puts function declarations into
alphabetic order in the header file.