Adjust CPU icon color according to frequency
This patch adds a new configuration option that, when enabled, replaces
black parts of the CPU icon with a color based on the current CPU
frequency. The minimum frequency leaves the icon unchanged, while the
maximum frequency fills the center of the icon with a red color.

The maximum frequency is being updated automatically based on the CPU
frequencies seen so far. This is more reliable than relying on the
maximum frequency reported by the Linux kernel via sysfs which does not
account for turbo frequencies or can be unrealistic.

It currently works only with the xfce4-cpufreq-plugin.png icon.

Closes: https://gitlab.xfce.org/panel-plugins/xfce4-cpufreq-plugin/-/issues/2
Co-authored-by: Gert Wollny
Co-authored-by: Andre Miranda <andreldm@xfce.org>