Improve static type-safety of signal handlers
This patch fixes the invalid return type of the label_draw() function.

The global CpuFreqPlugin variable now uses Ptr0<CpuFreqPlugin> instead of
CpuFreqPlugin*. This global variable should be removed in the future, in
order for the plugin instance to be deallocated automatically.
If xfce4-panel is started with the G_MESSAGES_DEBUG environment variable
set to "all": CpuFreqPlugin's destructor prints a console info message.
A similar message is printed in the case of CpuFreqPluginConfigure's
destructor.