Stop using xfce_panel_plugin_(block|unblock)_menu
It's counter-intuitive for the user to have the context menu entry for
displaying the plugin settings dialog grayed out when this dialog is
already shown, so it's better to leave this menu entry clickable and
bring the dialog back to the foreground if necessary. In addition, there
are other ways of triggering the display of this dialog, such as the
button in the panel settings dialog (items tab), whose state cannot be
synchronized with xfce_panel_plugin_(block|unblock)_menu.

Related: xfce/xfce4-panel#937